[MainRobot] Make robot connect & listen to camera.
This commit adds a CameraClient class: a separate thread to listen to the camera, and
perform basic computation on the markers.
Note that this code requires Eigen3 to compile.

Currently, the behavior of the client is only to count the elements in the excavation site,
and log it: this has no impact on the robot's strategy.